Restructured to seperate apps and templates
