+++ app/app_templates/__init__.py
@@ -1,0 +1,1 @@
+from app.app_templates.views import templates  # noqa
