+++ app/templates/app_templates/new_template.html
@@ -8,7 +8,7 @@
 {% block content %}
     <div class="ui stackable centered grid container">
         <div class="twelve wide column">
-            <a class="ui basic compact button" href="{{ url_for('apps.index') }}">
+            <a class="ui basic compact button" href="{{ url_for('templates.index') }}">
                 <i class="caret left icon"></i>
                 Back to dashboard
             </a>
