+++ app/templates/apps/view_apps.html
@@ -42,7 +42,7 @@
                     <thead>
                         <tr>
                             <th class="sorted ascending">App Name</th>
-                            <th>App ID</th>
+                            <th>App Ports</th>
                             <th>App Status</th>
                         </tr>
                     </thead>
@@ -51,8 +51,10 @@
                         <tr {# onclick="window.location.href = '{{ url_for('apps.template_info', template_id=t.id) }}';" #}>
                             <td>{{ a.name }}</td>
                             <td>
-                            {% for port in a.ports %}
-                                {{ port }}
+                            {% for port in a.ports.values() %}
+                                {% for entry in port %}
+                                    {{ entry }}
+                                {% endfor %}
                             {% endfor %}
                             </td>
                             <td>{{ a.status }}</td>
